k256: rename Schnorr `signing`/`verifying` modules
Renames modules as follows:

- `sign` => `signing`
- `verify` => `verifying`

This better matches the `SigningKey` and `VerifyingKey` type names.
